requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
anthropic==0.42.0
gspread==6.1.4
google-auth==2.37.0
pydantic==2.10.4
jinja2==3.1.5
